create a file/class per mcp client sample
